add sample code, readme and resolve comments
