+++ azure-keyvault-secret-spring-boot-starter-sample/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+azure.keyvault.client-id=put-your-azure-client-id-here
+azure.keyvault.client-key=put-your-azure-client-key-here
+azure.keyvault.uri=put-your-azure-keyvault-uri-here
+azure.keyvault.enabled=true
