+++ azure-keyvault-secret-spring-boot/src/main/java/com/microsoft/azure/keyvault/spring/boot/Constants.java
@@ -1,0 +1,9 @@
+package com.microsoft.azure.keyvault.spring.boot;
+
+public class Constants {
+    public static final String AZURE_KEYVAULT_VAULT_URI = "azure.keyvault.uri";
+    public static final String AZURE_CLIENTID = "azure.keyvault.client-id";
+    public static final String AZURE_CLIENTKEY = "azure.keyvault.client-key";
+    public static final String AZURE_KEYVAULT_ENABLED = "azure.keyvault.enabled";
+    public static final String AZURE_KEYVAULT_PROERTYSOURCE_NAME = "azurekv";
+}
