+++ azure-keyvault-secret-spring-boot/src/test/resources/application.properties
@@ -1,0 +1,4 @@
+azure.keyvault.client-id=testid
+azure.keyvault.client-key=testkey
+azure.keyvault.uri=fakeuri
+azure.keyvault.enabled=false
