--- azure-keyvault-spring-boot/pom.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>com.microsoft.azure</groupId>
-    <artifactId>azure-keyvault-spring-boot</artifactId>
-    <packaging>jar</packaging>
-
-    <parent>
-        <groupId>com.microsoft.azure</groupId>
-        <artifactId>azure-spring-boot-starter-parent</artifactId>
-        <version>0.1.3</version>
-        <relativePath>../azure-spring-boot-starter-parent/pom.xml</relativePath>
-    </parent>
-
-    <name>Azure KeyVault Spring Boot</name>
-    <description>Spring Boot package with Azure KeyVault as PropertySource</description>
-    <url>https://github.com/Microsoft/azure-spring-boot-starters</url>
-
-    <licenses>
-        <license>
-            <name>MIT</name>
-            <url>https://github.com/Microsoft/azure-spring-boot-starters/blob/master/LICENSE</url>
-            <distribution>repo</distribution>
-        </license>
-    </licenses>
-
-    <developers>
-        <developer>
-            <id>yungez</id>
-            <name>Yunge Zhu</name>
-            <email>yungez@microsoft.com</email>
-        </developer>
-    </developers>
-
-    <scm>
-        <connection>scm:git:git://github.com/Microsoft/azure-spring-boot-starters.git</connection>
-        <developerConnection>scm:git:ssh://github.com:Microsoft/azure-spring-boot-starters.git</developerConnection>
-        <url>https://github.com/Microsoft/azure-spring-boot-starters/tree/master</url>
-    </scm>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-test</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.microsoft.azure</groupId>
-            <artifactId>azure-keyvault</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.microsoft.aad</groupId>
-            <artifactId>adal4j</artifactId>
-        </dependency>
-    </dependencies>
-</project>
