--- azure-keyvault-spring-boot/src/main/java/com/microsoft/azure/keyvault/spring/boot/Constants.java
@@ -1,8 +1,0 @@
-package com.microsoft.azure.keyvault.spring.boot;
-
-public class Constants {
-    public static final String AZURE_KEYVAULT_VAULT_URI = "azure.keyvault.vaulturi";
-    public static final String AZURE_CLIENTID = "azure.keyvault.clientid";
-    public static final String AZURE_CLIENTKEY = "azure.keyvault.clientkey";
-    public static final String AZURE_KEYVAULT_ENABLED = "azure.keyvault.enabled";
-}
