--- azure-keyvault-spring-boot/src/test/resources/application.properties
@@ -1,4 +1,0 @@
-azure.keyvault.clientid=testid
-azure.keyvault.clientkey=testkey
-azure.keyvault.vaulturi=fakeuri
-azure.keyvault.enabled=true
