+++ azure-spring-boot-starter-parent/pom.xml
@@ -37,7 +37,7 @@
             <name>Zhijun Zhao</name>
             <email>zhijzhao@microsoft.com</email>
         </developer>
-        
+
         <developer>
             <id>yungez</id>
             <name>Yunge Zhu</name>
@@ -165,8 +165,9 @@
         <module>../azure-servicebus-spring-boot-starter-sample</module>
         <module>../spring-data-azure-documentdb</module>
         <module>../spring-data-azure-documentdb-sample</module>
-        <module>../azure-keyvault-spring-boot</module>
-        <module>../azure-keyvault-spring-boot-starter</module>
+        <module>../azure-keyvault-secret-spring-boot</module>
+        <module>../azure-keyvault-secret-spring-boot-starter</module>
+		<module>../azure-keyvault-secret-spring-boot-starter-sample</module>
         <module>../azure-mediaservices-spring-boot-autoconfigure</module>
         <module>../azure-mediaservices-spring-boot-starter</module>
         <module>../azure-mediaservices-spring-boot-starter-sample</module>
